moved functions to draw.c with header
