+++ src/quest.h
@@ -1,66 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-#include "crossplatform.h"
 #include <sys/time.h>
-#include "framebuffer.h"
-#include "menu.h"
+#include "crossplatform.h"
 #include <math.h>
+#include "framebuffer.h"
+#include "draw.h"
 
-struct buffer init_buffer();
-int map(int x, int inMin, int inMax, int outMin, int outMax);
-struct position kb_control(struct position player);
-void ray_cast(struct position player, unsigned short *distance);
-void draw_3d(unsigned short *distance, struct buffer fb);
-void draw_map(struct position player, struct buffer fb);
-void draw_menu(struct buffer fb);
-
-#define WORLDSIZE 20
-struct position{
-    float pX; // player position X
-    float pY; // player position Y
-    int pA; // player rotation in degree (180 = north)
-};
-struct level {
-    int world[WORLDSIZE][WORLDSIZE]; // birds perspective of world (world needs a boarder!)
-    struct position spawn;
-};
-
-struct level lvl = { // current level lvl
-    {
-        {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-    }, 
-    10, 10, 180
-};
-
-// menu
-unsigned short isMenu = 1;
-int loop = 1; // loop until exit
-
-const unsigned short mapS = 10; // minimap size // must be smaller then WORLDSIZE
-const short fov = 90; //set field of view to 90 degree
-const short maxVDist = 20; //set max viewing distran
-const float stepXY = 0.5; // player step size
-const int stepA = 5;  // player rotation steps
-int cliX = 144; // default fallback windows size 
-int cliY = 48;
-double cliA;  // width in degrees of one character in window size
